Fix compile error

+++ src/hash.cpp
@@ -168,7 +168,7 @@ typename enable_if<_Sz == 8, void>::type
 __check_for_overflow(size_t N)
 {
 #ifndef _LIBCPP_NO_EXCEPTIONS
-    if (N > 0xFFFFFFFFFFFFFFC5ull)
+    if ((unsigned long long)N > 0xFFFFFFFFFFFFFFC5ull)
         throw overflow_error("__next_prime overflow");
 #else
     (void)N;
